Add middleware to project
A middleware is used to handle a certain functionality. It can be chained
with other middlewares to form an execution chain. A middleware is
responsible to call the next middleware in the chain.

A middleware gets passed a context. The context is shared between all
middlewares of the same chain. Using the context, middlewares may
exchange data with other middlewares.