Make the context storage type-safe
To make the context type-safe, a new dedicated key for context values
is defined. This key contains an identifier and a type. When the same key
is used to store and retrieve a value from the context, type-safety
is guaranteed. Note, that the key overrides the equals and hashCode
methods, comparing the key's identifier and type. Therefor, it is not
necessary to use the same key instance when storing / retrieving values.

In addition to the improved context storage, this commit applies the
necessary changes to the already existing code base.