Implement AlpineBits router and routing middleware
An AlpineBits router should be capable to route AlpineBits requests
based on the X-AlpineBits-ClientProtocolVersion header (from here
on called version) and the action provided by the multipart/form-data
body.

The routing in its basic form tries to match a routing configuration
with the version and action. If a match is found, an middleware
mapped to the version and action should be invoked.

This commit adds the alpinebits-routing module, that provides the
necessary router, router configuration builder and middleware to
cover all the areas mentioned above:

- a Router interface, defining methods that router should implement,
  together with a default implementation
- a Builder to facilitate the route construction and configuration
- a Middleware that uses the router to decide, which middleware
  should be called next