Add alpinebits-housekeeping module
The module contains the two housekeeping middlewares for the
getVersion and getCapabilitites actions.

The middlewares look for a getVersion or getCapabilities action
in the context and add an according response value to the context.

Note, that this middlewares are not responsible to write the
result e.g. into a HttpServletResponse. This has to be done
somewhere else. Splitting up the duties this way has the advantage,
that the two middlewares are agnostic about the underlaying
communication infrastructure.