Implement mapping module
The mapping module provides facilities and abstractions
to simplify the work with AlpineBits data.

It provides abstracter data classes, that are not tied
to different AlpineBits versions. In addition, mappers
are implemented, that facilitate the conversion between
low-level AlpineBits data classes and the more abstracter
ones. Finally, the implemented middlewares can be
used to perform the mapping as part of ordinary middleare
chains.