Add example for GuestRequests with DB persistence
The example implements an AlpineBits server with support for
AlpineBits Housekeeping and GuestRequests actions. The persistence
is implemented using an in-memory H2 database.

Run the following command to build and start the server:

mvn clean install cargo:run

When the example is started using the command above,
a Tomcat 8 server boots up and deploys the AlpineBits server
as application with path /examples-dbserver.

The application provides an AlpineBits servlet as entrypoint
to the server. It is available at /examples-dbserver/AlpineBits.
You can POST to it.

The application provides also a simple GUI at
/examples-dbserver/index.html, that can be used to perform
pre-configured GuestRequests against the server.