Implement OTA extensions
The AlpineBits inventory actions specify, that any OTA
element/attribute matching the XSD can be used at the following
positions as an extension:

- OTAHotelDescriptiveContentNotifRQ -> HotelDescriptiveContents ->
HotelDescriptiveContent -> AffiliationInfo
- OTAHotelDescriptiveContentNotifRQ -> HotelDescriptiveContents ->
HotelDescriptiveContent -> ContactInfos
- OTAHotelDescriptiveContentNotifRQ -> HotelDescriptiveContents ->
HotelDescriptiveContent -> HotelInfo
- OTAHotelDescriptiveContentNotifRQ -> HotelDescriptiveContents ->
HotelDescriptiveContent -> Policies

This commit provides a module to simplify the usage of the OTA
extensions. In addition, it extends the existing mapping
functionality to handle OTA extensions.