Add full XML conversion tests for AlpineBits 2018-10
This test converts the XML samples provided by the AlpineBits alliance
into Java objects, converts those objects back to XML and then compares
the original and the resulting XML.

Please note, that it was necessary to modify the sample
GuestRequests-OTA_ResRetrieveRS-request-with-roomtype.xml, adding
a zone to the datetime value of the attribute "CreateDateTime".
The original sample's value was "2017-09-03T19:47:50", now it is
"2017-09-03T19:47:50+01:00"

This modification was necessary, because the conversion from Java object
to XML for datetime values always adds the zone info. In this case,
the original the resulting XML wouldn't be equal.