Implement JAXB context provider for lazy initialisation and reuse of the JAXBContext.
JAXB context creation is a resource intensive operation. Since there
are a lot of classes involved in the OTA extension (basically the
whole OTA 2015), it takes quite some time to initialize the JAXB
context.

The implemented provider does a lazy initialization of the JAXB
context when the the class is loaded and returns the initialized
context on invocations of getJAXBContext().