Use OTA-2015A ObjectFactory for JAXB conversion facilities
Using the ObjectFactory when creating the JAXBContext fixes problems
where some classes were not serialized/deserialized correctly. This
also makes the registration hack introduced in commit c1dc358ae68d
superfluous.

In addition, the JAXBContext is now lazy-loaded and shared using a
singleton. This reduces startup and test times by a large amount.

The downside is, that JAXBXmlToObjectConverter and
JAXBObjectToXmlConverter can now only be used with OTA-2015A classes
but that shouldn't be a problem, because their sole purpose is to
facilitate the serialization/deserialization of OTA-2015A classes.