Update examples to use OTA-2015A base classes
The examples for Housekeeping, FreeRooms and Inventory are updated
and use now the OTA-2015A base classes.

The GuestRequest example was removed. It would have been a lot of
effort to convert it, because its implementation was largely based on
the alpinebits-mapping module, that was also removed. Maybe, the
GuestRequest example returns in a future commit.