Add AlpineBits 2020-10 samples and XML conversion tests and enhance XSD
The given XSD lacked definitions for the 2020-10 Inventory action. To
get a to a working system and to make all 2020-10 test run, the
definitions were added manually. Be aware, that those definitions are
not the ones from the final 2020-10 standard! As soon as the standard
is complete, the XSD definitions will be taken from there.