Fix two example files to match expected values
The example files are used in full-conversion tests, i.e. tests that convert an XML
file to objects and then the objects to XML. Finally, the two XML files are compared.
The tests fail if differences are found.

Due to the implementation detail of Java date-time serialization and the configuration
used, the date-time values are serialized with additional date-time offsets (allowed by
AlpineBits). Those offsets are not part of official the examples (also allowed by
AlpineBits), therefore the tests fail.

To prevent the tests from failing, the date-time offset is added to date-time values
in the examples.