Add OTA 2015a XSD validation
This commit adds the OTA 2015a XSD validation (as required by the AlpineBits standard)
in addition to the already existing AlpineBits XSD validation.

The added validation requires the XML input / output streams to be stored into
temporary strings, because now the XML content has to be reused several times,
which would fail:

- for OTA 2015a XSD validation
- for AlpineBits XSD validation and XML conversion using JAXB

Although JAXB is capable of XSD validation while XML conversion (this is the way
the AlpineBits XSD validation is handled at the moment), JAXB is restricted to a
single XSD schema. This complicates the validation against different schemas (like
OTA 2015a XSD and AlpineBits XSD) and necessitates an additional XSD validation step.

Both the additional XSD validation step and JAXB consume provided streams, therefore
it was necessary to introduce temporary strings that can be handed to several parties.

In addtion to the implementation, some tests had to be adjusted, because the additional
XSD validation catches some errors that were handled previousely by business validation
rules.